fistaszek/requirements.txt
Babel==2.3.4
Django==1.8.12
Jinja2==2.8
MarkupSafe==0.23
Morelia==0.6.1
Pygments==2.1.3
Sphinx==1.4.1
alabaster==0.7.7
amqp==1.4.9
anyjson==0.3.3
argparse==1.2.1
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
celery==3.1.23
coverage==4.1b2
decorator==4.0.9
distribute==0.7.3
django-celery==3.1.17
django-extensions==1.6.1
django-shell-plus==1.1.7
django-smarttest==0.3.0
docutils==0.12
factory-boy==2.7.0
fake-factory==0.5.7
funcsigs==1.0.0
gunicorn==19.4.5
imagesize==0.7.0
ipaddress==1.0.16
ipdb==0.9.3
ipython==4.2.0
ipython-genutils==0.1.0
kombu==3.0.35
ordereddict==1.1
parse==1.6.6
pbr==1.9.1
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.7.2
psycopg2==2.6.1
ptyprocess==0.5.1
python-dateutil==2.5.3
pytz==2016.4
q==2.6
redis==2.10.5
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
traitlets==4.2.1
wsgiref==0.1.2
mock==2.0.0